[Feat/#24] 파일 분리
